# edge/medbook-bff/src/main/resources/openapi/bff-api_v1.0.yaml
openapi: 3.0.3
...
info:
  title: MedBook - BFF API
  description: |
    ## MedBook BFF
    Unico punto di ingresso del frontend verso il sistema.
    Orchestra le chiamate ai DMN via Feign client e restituisce risposte aggregate.
    Non ha database proprio. E' un orchestratore stateless.

    ### Ruoli supportati
    - ROLE_PATIENT: paziente registrato
    - ROLE_DOCTOR: medico
    - ROLE_RECEPTIONIST: addetto alla reception
    - ROLE_ADMIN: amministratore

  version: 1.0.0
  contact:
    name: MedBook Team
    email: djizalfra@gmail.com
servers:
  - url: http://localhost:8090
    description: BFF locale - sviluppo
...
paths:
...
  /bff/v1/doctors/me:

    get:
      tags:
        - Doctors
      operationId: getMyDoctor
      summary: Profilo medico dell'utente autenticato
      description: |
        Restituisce il profilo del medico autenticato.
        Il doctorId viene risolto internamente cercando per lo username (CF per pazienti, email per medici) del claim
        'preferred_username' nel token JWT - non e' necessario passarlo come parametro.
        Il risultato e' cached in L1 (request scope) e L2 (Caffeine, TTL 60 min).
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      responses:
        '200':
          description: Profilo medico trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '404':
          description: Profilo medico non trovato per l'utente autenticato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
  /bff/v1/doctors:

    post:
      tags:
        - Doctors
      operationId: postCreateDoctor
      summary: Crea un nuovo medico
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDoctorBffRequest'
      responses:
        '201':
          description: Medico creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '422':
          description: Dati non validi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    get:
      tags:
        - Doctors
      operationId: getAllDoctors
      summary: Lista medici con filtri
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: lastName
          in: query
          required: false
          schema:
            type: string
        - name: specialization
          in: query
          required: false
          schema:
            type: string
        - name: firstName
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: phone
          in: query
          required: false
          schema:
            type: string
        - name: licenseNumber
          in: query
          required: false
          schema:
            type: string
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: updatedFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: updatedTo
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista medici recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
  /bff/v1/doctors/{doctorId}:

    get:
      tags:
        - Doctors
      operationId: getDoctorById
      summary: Dettaglio medico
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
      responses:
        '200':
          description: Medico trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '404':
          description: Medico non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    patch:
      tags:
        - Doctors
      operationId: patchUpdateDoctor
      summary: Aggiorna dati medico
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDoctorBffRequest'
      responses:
        '200':
          description: Medico aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'

    delete:
      tags:
        - Doctors
      operationId: deleteDoctor
      summary: Elimina logicamente un medico
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
      responses:
        '200':
          description: Medico eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
  /bff/v1/doctors/{doctorId}/restore:

    patch:
      tags:
        - Doctors
      operationId: patchRestoreDoctor
      summary: Ripristina un medico eliminato logicamente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
      responses:
        '200':
          description: Medico ripristinato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Medico non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
        '409':
          description: Il medico non e cancellato - impossibile ripristinare
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

  # ===========================================================================
  # SPECIALIZATIONS
  # ===========================================================================
  /bff/v1/specializations:

    get:
      tags:
        - Specializations
      operationId: getSpecializations
      summary: Lista specializzazioni disponibili
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      responses:
        '200':
          description: Lista specializzazioni recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'

  # ===========================================================================
  # AVAILABILITIES (template per medico)
  # ===========================================================================
  /bff/v1/doctors/{doctorId}/availabilities:

    post:
      tags:
        - Availabilities
      operationId: postCreateAvailability
      summary: Aggiunge template di disponibilita al medico
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAvailabilityBffRequest'
      responses:
        '201':
          description: Template creato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '409':
          description: Template gia esistente per questa combinazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    get:
      tags:
        - Availabilities
      operationId: getAllAvailabilities
      summary: Lista template disponibilita del medico
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
      responses:
        '200':
          description: Lista template recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
  /bff/v1/doctors/{doctorId}/availabilities/{clinicId}/{dayOfWeek}/{startTime}:

    delete:
      tags:
        - Availabilities
      operationId: deleteAvailability
      summary: Rimuove logicamente un template di disponibilita
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
        - name: dayOfWeek
          in: path
          required: true
          schema:
            type: string
        - name: startTime
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Template non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
  /bff/v1/doctors/{doctorId}/availabilities/{clinicId}/{dayOfWeek}/{startTime}/restore:

    patch:
      tags:
        - Availabilities
      operationId: patchRestoreAvailability
      summary: Ripristina un template di disponibilita eliminato logicamente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/DoctorIdPathParam'
        - name: clinicId
          in: path
          required: true
          schema:
            type: string
        - name: dayOfWeek
          in: path
          required: true
          schema:
            type: string
        - name: startTime
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template ripristinato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Template non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
        '409':
          description: Il template non e cancellato - impossibile ripristinare
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

  # ===========================================================================
  # CLINICS
  # ===========================================================================
...
components:
...
  parameters:
    MedBookContextParam:
      name: X-MedBook-Context
      in: header
      required: false
      description: |
        Contesto applicativo della richiesta - iniettato da MedBookContextInterceptor.
        Non viene letto dall'header HTTP: il resolver preleva MedBookContextHolder.get().
      schema:
        $ref: '#/components/schemas/MedBookContext'
...
    DoctorIdPathParam:
      name: doctorId
      in: path
      required: true
      schema:
        type: string
        example: DOC-1
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: createdAt,desc

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    MedBookApiResponse:
      type: object
      description: Risposta di successo con payload.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        data:
          description: Payload della risposta
        page:
          $ref: '#/components/schemas/MedBookPageResponse'
    MedBookApiVoidResponse:
      type: object
      description: Risposta base per endpoint senza payload.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
    MedBookApiErrorResponse:
      type: object
      description: Formato standard delle risposte di errore.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        errorCode:
          type: string
        message:
          description: Messaggio di errore
    MedBookPageResponse:
      type: object
      description: Metadati di paginazione.
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    MedBookContext:
      type: object
      description: Contesto della richiesta HTTP.
      properties:
        traceId:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        requestTimestamp:
          type: string
          format: date-time

    # =========================================================================
    # PATIENT REQUEST
    # =========================================================================
...
    CreateDoctorBffRequest:
      type: object
      description: Dati per la creazione di un nuovo medico
      required:
        - firstName
        - lastName
        - email
        - phone
        - licenseNumber
      properties:
        firstName:
          type: string
          example: Luca
        lastName:
          type: string
          example: Bianchi
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          example: MASCHILE
        email:
          type: string
          format: email
          example: luca.bianchi@medbook.it
        phone:
          type: string
          example: "+393331234567"
        licenseNumber:
          type: string
          example: RM-12345
        emailEnabled:
          type: boolean
          description: true se il canale EMAIL è abilitato
          example: true
        smsEnabled:
          type: boolean
          description: true se il canale SMS è abilitato
          example: false
    UpdateDoctorBffRequest:
      type: object
      description: Dati per l'aggiornamento parziale di un medico
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string

    # =========================================================================
    # NOTIFICATION PREFERENCES REQUEST
    # =========================================================================
...
    CreateAvailabilityBffRequest:
      type: object
      description: Template di disponibilita settimanale da creare per il medico
      required:
        - availabilities
      properties:
        availabilities:
          type: array
          items:
            $ref: '#/components/schemas/AvailabilityItemBff'
    AvailabilityItemBff:
      type: object
      description: Singolo template di disponibilita da creare
      required:
        - clinicId
        - dayOfWeek
        - startTime
        - endTime
      properties:
        clinicId:
          type: string
          example: CLN-1
        dayOfWeek:
          type: string
          example: LUNEDI
        startTime:
          type: string
          example: "09:00"
        endTime:
          type: string
          example: "13:00"

    # =========================================================================
    # CLINIC REQUEST
    # =========================================================================
...